added test for logging in via room, moved parent to another package
