+++ src/test/java/LiveChatAPI/LiveChatRoomTest/CloseConversationTest.java
@@ -1,5 +1,6 @@
 package LiveChatAPI.LiveChatRoomTest;
 
+import LiveChatAPI.LiveChatRoomTest.ChatRoomParent.RoomParent;
 import io.rocketchat.common.data.model.ErrorObject;
 import io.rocketchat.livechat.callback.MessageListener;
 import io.rocketchat.livechat.model.GuestObject;
@@ -17,7 +18,7 @@ import static org.mockito.Mockito.timeout;
 /**
  * Created by sachin on 17/7/17.
  */
-public class CloseConversationTest  extends RoomParent{
+public class CloseConversationTest  extends RoomParent {
 
     @Mock
     MessageListener.SubscriptionListener subscriptionListener;
