+++ src/test/java/LiveChatAPI/LiveChatRoomTest/SubscribeTest.java
@@ -1,5 +1,6 @@
 package LiveChatAPI.LiveChatRoomTest;
 
+import LiveChatAPI.LiveChatRoomTest.ChatRoomParent.RoomParent;
 import io.rocketchat.common.data.model.ErrorObject;
 import io.rocketchat.livechat.model.GuestObject;
 import org.junit.After;
@@ -8,7 +9,7 @@ import org.junit.Before;
 /**
  * Created by sachin on 17/7/17.
  */
-public class SubscribeTest extends RoomParent{
+public class SubscribeTest extends RoomParent {
 
     @Before
     public void setup(){
